Changes list of children in chapter children component with item chapter list component
